sampling/pyproject.toml
[project]
name = "sampling"
version = "0.1.0"
description = "Demonstration of sampling with MCP"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aioconsole>=0.8.1",
    "mcp[cli]>=1.9.3",
    "openai==2.30.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]

[tool.setuptools]
py-modules = ["client", "server"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
